Camera Boundry and Updated Player Movement
Switched from Fixed update to a method for player movement and multiplying by time.delta time instead to ensure independence from frame rate. Additionally added a variable intermediary variable for the new position which we clamp to fit inside of the camera boundry which in turn is obtained from a new function that initializes the camera bounds using Viewport.