Updated WaveConfigSO with most of the code
Made a scriptable object with functions that allow us to get both the starting waypoint and the entire list of waypoints. Additionally I'm setting up the future enemy speed inside of this scriptable object and added a function to get the move speed. I did make these functions public since they act as getters, I still don't know if this is ideal so I might come back to change that.