Setup Enemy Spawner + added Pathfinder to Enemy Prefab
Since enemies are instantiated now through the enemySpawner I removed the enemy I had previously from the inspector, as enemies are now stored in each WaveConfig. I also realized my enemyPrefab didn't have the pathfinding script so I just added that in. Everything works fine now, although the inspector might get cluttered with enemies in the future so I'm working on a fix for that