Created Health and Damage Script
Both need a trigger collider on the gameObject to work. Damage Script simply defines how much damage is done by an object. Health Script tracks collisions and checks if damage script is exitent on the gameObject collided with, if it is then it simply gets the damage amount from the damage script and substracts it from the health.  It also destroys the game object when health is < 0, as does the damage dealer destroy itself when hitting something.