HealthFix + Editor Changes
Fixed a bug caused by calling the play() method from the camera shake script directly from health instead of using thee shakeCamera() method from within the health script that checks if applyCameraShakeOnDamageTaken is true. In addition I toggled the bool to be set to true inside of the Unity Editor for the player prefab.