Created ScoreKeeper + Updated Health
Created a scoreKeeper script that keeps track of the score and contains a few methods to modify, reset, and get the score. Created a definition of scorekeeper within the health script and added a new variable for the amount of points given on death for any given entity. If not AI points are added through a reference to the ModifyScore() method in ScoreKeeper