ScoreKeeperUpdate + Health Fix
Added a reference to UIDisplay in scorekeeper so that it can update the score. Also, fixed a bug where I was adding points only if the player died (which would give 0 points) instead of when the enemy dies. This was caused by the ! (not symbol) being written by my sleep-deprived self after the boolean instead of before, smh.