Painstaking BugFix: Score updates after reset now
Took an hour of debugging to realize what was going on. The scorekeeper script was trying to still reference the old UIDisplay, which of course no longer existed, yet Unity decided to give no errors in the console for this, so it took a century to find out that's what was happening. Thankfully it was simple to fix by just calling a reload method once UIDisplay has been created. I also fixed a minor bug with the score not resetting but I lowkey just added another reference to scorekeeper once in the LoadGame method inside of level manager and that seemed to fix it.