Leaderboard Script
I'm not going to lie, I have no idea what I just did. I followed documentation for this and I essentially just get the names and scores from the leaderboard and assign them to an array of names and scores in my unity scene. Uploading the new entry is a similar process where I just pass in a username and a score, what I do get are the callback functions which I have yet to learn but are needed here. I'll have to revisit that at some point to fully understand